backend/pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [ "Marcus Schiesser <mail@marcusschiesser.de>" ]
readme = "README.md"

[tool.poetry.scripts]
generate = "app.engine.generate:generate_datasource"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.109.1"
python-dotenv = "^1.0.0"
aiostream = "^0.5.2"
llama-index = "0.10.58"
cachetools = "^5.3.3"
pdf2image = "1.17.0"
pytesseract = "0.3.10"

llama-index-core = "0.10.58"
llama-index-llms-ollama = "0.1.5"
llama-index-embeddings-ollama = "0.1.2"
llama-index-vector-stores-postgres = "0.1.13"
llama-index-readers-s3 = "<0.2.0"
llama-index-storage-chat-store-redis = "<0.2.0"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard" ]
version = "^0.23.2"

[tool.poetry.dependencies.docx2txt]
version = "^0.8"

[tool.poetry.dependencies.llama-index-agent-openai]
version = "0.2.6"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"